refactor(api): rename testing package to iristest, move CLI to cmd/iris
Move the public SDK test-utility package from
github.com/petal-labs/iris/testing to github.com/petal-labs/iris/iristest
so it no longer collides with the standard-library testing package, and
move the CLI entrypoint from cli/cmd/iris to cmd/iris alongside
cmd/gen-models.

The old testing import path is removed rather than shimmed, because a
forwarding package would preserve the naming collision. Test-double APIs,
CLI flags, output, and release artifact names are unchanged. Reusable CLI
packages stay under cli/commands, cli/config, and cli/keystore, so
version injection via -ldflags is unaffected.

Updates README install guidance, Makefile, integration-test binary build,
development docs, testing guide, the mock-provider example, and the
release workflow to the new paths.

Closes #102
