+++ iristest/mock_provider.go
@@ -1,4 +1,4 @@
-package testing
+package iristest
 
 import (
 	"context"
