+++ iristest/recording_provider_test.go
@@ -1,4 +1,4 @@
-package testing
+package iristest
 
 import (
 	"context"
